Create Roots_Quadratic_Equation
This way is to find out the roots using NumPy's .roots() the method. You can calculate it in a single line. This function can solve for any degree polynomial and since you will solve a quadratic equation, this function will take 3 arguments which represent the coefficients of the powers of x.